Discovered Webpack couldn't reach nested LabelMask components in Storybook build - moved files up a level.
