+++ packages/design-system/src/components/TextField/LabelMask.test.tsx
