+++ packages/design-system/src/components/TextField/LabelMask.tsx
