+++ packages/design-system/src/components/TextField/index.ts
@@ -1,4 +1,4 @@
 export { default as TextField } from './TextField';
 export { default as TextInput } from './TextInput';
 export { maskValue, unmaskValue } from './maskHelpers';
-export { DATE_MASK } from './LabelMask/LabelMask';
+export { DATE_MASK } from './LabelMask';
